Purges all traces of that blasted library named Armadillo (#428)
* Docker and image removals

* Removes eigen_armadillo support file

* Removes findArmadillo cmake file

* Removes arma includes

* Removes obsolete matrix utilities

* Removes UnitQuaternion

* Removes old matrix utility includes

* Removes proto_transform

* Removes some utility funcs

* Removes some cmake refs

* More utility funcs

* Deletes arma particle filters

* Removes arma quad version

* Removes obsolete matrix utility

* Removes arma forwardkinematics stuff

* Removes arma inverse kinematics

* Removes inversekinematics arma stuff on the hpp

* Removes armadillo support in yaml_expression

* Last traces

* Adds back mistakenly removed transform utility

* Fixes dead includes

* Fixes typos from last commit

* Final fixes

* Moves eigen particle filters

* Formatting fixes post merge

* Removes sneaky arma functions from merge

* Removes other sneaky arma functions from merge

* Removes sneaky transform3ds from merge

* FORMATTING PRS ARE OFFICIALLY THE WORST THING EVER

* Removes template specialisation for quad

* Formatting commits really really stink
