+++ module/behaviour/planning/KickPlanner/src/KickPlanner.cpp
@@ -63,7 +63,6 @@ namespace module::behaviour::planning {
     using LimbID = utility::input::LimbID;
     using utility::localisation::fieldStateToTransform3D;
     using utility::math::coordinates::sphericalToCartesian;
-    using utility::motion::kinematics::legPoseValid;
     using utility::nusight::graph;
 
     KickPlanner::KickPlanner(std::unique_ptr<NUClear::Environment> environment)
