+++ module/behaviour/skills/HeadBehaviourSoccer/src/HeadBehaviourSoccer.hpp
@@ -20,7 +20,6 @@
 #ifndef MODULES_BEHAVIOUR_REFLEX_HEADBEHAVIOURSOCCER_HPP
 #define MODULES_BEHAVIOUR_REFLEX_HEADBEHAVIOURSOCCER_HPP
 
-#include <armadillo>
 #include <nuclear>
 #include <set>
 
@@ -96,8 +95,8 @@ namespace module::behaviour::skills {
 
         /*! @brief Gets a bounding box in screen angular space of a set of vision objects
          */
-        utility::math::geometry::Quad<Eigen::Vector2d> getScreenAngularBoundingBox(const message::vision::Balls& obs);
-        utility::math::geometry::Quad<Eigen::Vector2d> getScreenAngularBoundingBox(const message::vision::Goals& obs);
+        utility::math::geometry::Quad<double, 2, 1> getScreenAngularBoundingBox(const message::vision::Balls& obs);
+        utility::math::geometry::Quad<double, 2, 1> getScreenAngularBoundingBox(const message::vision::Goals& obs);
 
         bool orientationHasChanged(const message::input::Sensors& sensors);
