+++ module/localisation/RobotParticleLocalisation/src/RobotParticleLocalisation.hpp
@@ -10,7 +10,7 @@
 #include "message/support/FieldDescription.hpp"
 #include "message/vision/Goal.hpp"
 
-#include "utility/math/filter/eigen/ParticleFilter.hpp"
+#include "utility/math/filter/ParticleFilter.hpp"
 
 namespace module::localisation {
