+++ module/platform/darwin/SensorFilter/src/SensorFilter.cpp
@@ -27,7 +27,6 @@
 
 #include "utility/input/LimbID.hpp"
 #include "utility/input/ServoID.hpp"
-#include "utility/math/matrix/matrix.hpp"
 #include "utility/motion/ForwardKinematics.hpp"
 #include "utility/nusight/NUhelpers.hpp"
 #include "utility/platform/darwin/DarwinSensors.hpp"
