+++ module/platform/darwin/SensorFilter/src/SensorFilter.hpp
@@ -29,7 +29,7 @@
 
 #include "message/motion/KinematicsModel.hpp"
 
-#include "utility/math/filter/eigen/UKF.hpp"
+#include "utility/math/filter/UKF.hpp"
 
 
 namespace module::platform::darwin {
