+++ module/vision/BallDetector/src/BallDetector.cpp
@@ -21,6 +21,7 @@
 
 #include <fmt/format.h>
 #include <fmt/ostream.h>
+#include <numeric>
 
 #include "extension/Configuration.hpp"
