+++ module/vision/GoalDetector/src/GoalDetector.cpp
@@ -22,6 +22,7 @@
 #include <Eigen/Core>
 #include <Eigen/Geometry>
 #include <fmt/format.h>
+#include <numeric>
 
 #include "extension/Configuration.hpp"
