+++ shared/utility/libraries.cmake
@@ -13,9 +13,6 @@ target_link_libraries(nuclear_utility PUBLIC fmt::fmt)
 find_package(Aravis REQUIRED)
 target_link_libraries(nuclear_utility PUBLIC Aravis::Aravis)
 
-find_package(Armadillo REQUIRED)
-target_link_libraries(nuclear_utility PUBLIC Armadillo::Armadillo)
-
 if(CMAKE_BUILD_TYPE STREQUAL "Debug")
   find_package(libbacktrace REQUIRED)
   target_link_libraries(nuclear_utility PUBLIC libbacktrace::libbacktrace ${CMAKE_DL_LIBS})
