+++ shared/utility/math/coordinates.hpp
@@ -21,7 +21,6 @@
 #define UTILITY_MATH_COORDINATES_HPP
 
 #include <Eigen/Core>
-#include <armadillo>
 #include <cmath>
 
 
@@ -33,20 +32,6 @@
  * @author Alex Biddulph
  */
 namespace utility::math::coordinates {
-    inline arma::vec3 sphericalToCartesian(const arma::vec3& sphericalCoordinates) {
-        double distance  = sphericalCoordinates[0];
-        double cos_theta = cos(sphericalCoordinates[1]);
-        double sin_theta = sin(sphericalCoordinates[1]);
-        double cos_phi   = cos(sphericalCoordinates[2]);
-        double sin_phi   = sin(sphericalCoordinates[2]);
-        arma::vec3 result;
-
-        result[0] = distance * cos_theta * sin_phi;
-        result[1] = distance * sin_theta * sin_phi;
-        result[2] = distance * cos_phi;
-
-        return result;
-    }
 
     template <typename T, typename U = typename T::Scalar>
     inline Eigen::Matrix<U, 3, 1> sphericalToCartesian(const Eigen::MatrixBase<T>& sphericalCoordinates) {
@@ -60,24 +45,6 @@ namespace utility::math::coordinates {
         result.x() = distance * cos_theta * cos_phi;
         result.y() = distance * sin_theta * cos_phi;
         result.z() = distance * sin_phi;
-
-        return result;
-    }
-
-    inline arma::vec3 cartesianToSpherical(const arma::vec3& cartesianCoordinates) {
-        double x = cartesianCoordinates[0];
-        double y = cartesianCoordinates[1];
-        double z = cartesianCoordinates[2];
-        arma::vec3 result;
-
-        result[0] = sqrt(x * x + y * y + z * z);  // r
-        result[1] = atan2(y, x);                  // theta
-        if (result[0] == 0) {
-            result[2] = 0;
-        }
-        else {
-            result[2] = asin(z / (result[0]));  // phi
-        }
 
         return result;
     }
@@ -101,35 +68,6 @@ namespace utility::math::coordinates {
         return result;
     }
 
-    inline arma::vec4 sphericalToCartesian4(const arma::vec3& sphericalCoordinates) {
-        arma::vec3 p = sphericalToCartesian(sphericalCoordinates);
-        return arma::vec4({p[0], p[1], p[2], 1});
-    }
-
-    inline arma::vec4 cartesianToSpherical4(const arma::vec3& cartesianCoordinates) {
-        arma::vec3 p = cartesianToSpherical(cartesianCoordinates);
-        return arma::vec4({p[0], p[1], p[2], 1});
-    }
-
-
-    inline arma::vec2 cartesianToRadial(const arma::vec2& cartesianCoordinates) {
-        double x = cartesianCoordinates[0];
-        double y = cartesianCoordinates[1];
-
-        return {sqrt(x * x + y * y), atan2(y, x)};
-    }
-
-    inline arma::vec2 spherical2Radial(const arma::vec3& sphericalCoordinates) {
-        double dist          = sphericalCoordinates(0);
-        double declination   = sphericalCoordinates(2);
-        double flat_distance = dist * std::cos(declination);
-
-        arma::vec2 result;
-        result[0] = flat_distance;
-        result[1] = sphericalCoordinates(1);
-
-        return result;
-    }
 }  // namespace utility::math::coordinates
