+++ shared/utility/math/geometry/Quad.hpp
@@ -21,209 +21,13 @@
 #define UTILITY_MATH_GEOMETRY_QUAD_HPP
 
 #include <Eigen/Core>
-#include <armadillo>
 #include <ostream>
 #include <vector>
 
 namespace utility::math::geometry {
 
-    template <typename T>
-    class Quad : std::false_type {};
-
-    // *********************
-    // * ARMADILLO VERSION *
-    // *********************
-
-    template <typename Scalar>
-    class Quad<typename arma::Col<Scalar>> {
-    public:
-        using T = typename arma::Col<Scalar>::template fixed<2>;
-        Quad() : bl(arma::fill::zeros), br(arma::fill::zeros), tr(arma::fill::zeros), tl(arma::fill::zeros) {}
-        Quad(const Quad& other) : bl(other.bl), br(other.br), tr(other.tr), tl(other.tl) {}
-        Quad(const T& bottomLeft, const T& topLeft, const T& topRight, const T& bottomRight)
-            : bl(bottomLeft), br(bottomRight), tr(topRight), tl(topLeft) {}
-        Quad(const Scalar& left, const Scalar& top, const Scalar& right, const Scalar& bottom)
-            : bl(left, bottom), br(right, bottom), tr(right, top), tl(left, top) {}
-
-        /**
-         * Sets the Quad as a screen aligned rectangle given the specified positions.
-         * @param left     The left x pixel value.
-         * @param top      The top y pixel value.
-         * @param right    The right x pixel value.
-         * @param bottom   The bottom y pixel value.
-         */
-        void set(const Scalar& left, const Scalar& top, const Scalar& right, const Scalar& bottom) {
-            bl.x() = left;
-            bl.y() = bottom;
-            br.x() = right;
-            br.y() = bottom;
-            tl.x() = left;
-            tl.y() = top;
-            tr.x() = right;
-            tr.y() = top;
-        }
-
-        /**
-         * Sets the Quad given the specified corners.
-         * @param bottomLeft  The bottom left corner.
-         * @param topLeft     The top left corner.
-         * @param topRight    The top right corner.
-         * @param bottomRight The bottom right corner.
-         */
-        void set(const T& bottomLeft, const T& topLeft, const T& topRight, const T& bottomRight) {
-            bl = bottomLeft;
-            tl = topLeft;
-            tr = topRight;
-            br = bottomRight;
-        }
-
-        //! Returns the bottom centre pixel location of the Quad.
-        T getTopCentre() const {
-            return ((tl + tr) * 0.5);
-        }
-        //! Returns the bottom centre pixel location of the Quad.
-        T getBottomCentre() const {
-            return ((bl + br) * 0.5);
-        }
-        T getRightCentre() const {
-            return ((br + tr) * 0.5);
-        }
-        T getLeftCentre() const {
-            return ((bl + tl) * 0.5);
-        }
-
-        //! Returns the centre pixel location  of the Quad.
-        T getCentre() const {
-            return ((bl + tl + tr + br) * 0.25);
-        }
-
-        //! Returns the bottom left pixel location  of the Quad.
-        T getBottomLeft() const {
-            return bl;
-        }
-        //! Returns the bottom right pixel location  of the Quad.
-        T getBottomRight() const {
-            return br;
-        }
-        //! Returns the top left pixel location  of the Quad.
-        T getTopLeft() const {
-            return tl;
-        }
-        //! Returns the top right pixel location  of the Quad.
-        T getTopRight() const {
-            return tr;
-        }
-
-        // Returns the bounding box width and height
-        T getSize() const {
-            Quad boundingBox = getBoundingBox(getVertices());
-            return T({boundingBox.getAverageWidth(), boundingBox.getAverageHeight()});
-        }
-
-        Scalar getLeft() const {
-            return (0.5 * (bl(0) + tl(0)));
-        }
-        Scalar getRight() const {
-            return (0.5 * (br(0) + tr(0)));
-        }
-        Scalar getTop() const {
-            return (0.5 * (tl(1) + tr(1)));
-        }
-        Scalar getBottom() const {
-            return (0.5 * (bl(1) + br(1)));
-        }
-
-        // //! Returns the base width of the Quad in pixels.
-        // int getBaseWidth() const;
-        // //! Returns the top width of the Quad in pixels.
-        // int getTopWidth() const;
-
-        // //! Returns the left height of the Quad in pixels.
-        // int getLeftHeight() const;
-        // //! Returns the right height of the Quad in pixels.
-        // int getRightHeight() const;
-
-        //! Returns the average width of the Quad in pixels.
-        Scalar getAverageWidth() const {
-            return ((0.5 * (arma::norm(br - bl) + arma::norm(tr - tl))));
-        }
-
-        //! Returns the average height of the Quad in pixels.
-        Scalar getAverageHeight() const {
-            return ((0.5 * (arma::norm(br - tr) + arma::norm(bl - tl))));
-        }
-
-        Scalar area() const {
-            // Area of a quadrilateral: A = 0.5 * |diag1 X diag2|
-            // In two dimensions, this equates to: A = 0.5 * |(diag1.x)(diag2.y) - (diag2.x)(diag2.y)|
-            T diag1 = bl - tr;
-            T diag2 = tl - br;
-            return std::abs(0.5 * ((diag1(0) * diag2(1)) - (diag1(1) * diag2(0))));
-        }
-        Scalar aspectRatio() const {
-            return ((arma::norm(br - tr) + arma::norm(bl - tl) + 2) / (arma::norm(br - bl) + arma::norm(tr - tl) + 2));
-        }
-
-        std::vector<T> getVertices() const {
-            std::vector<T> vert = {tr, br, bl, tl};
-            return vert;
-        }
-
-        bool overlapsHorizontally(const Quad& other) const {
-            // Rough for now.
-            Scalar farRight   = std::max(tr(0), br(0));
-            Scalar farLeft    = std::min(tl(0), bl(0));
-            Scalar o_farRight = std::max(other.tr(0), other.br(0));
-            Scalar o_farLeft  = std::min(other.tl(0), other.bl(0));
-
-            return !((farRight < o_farLeft) || (o_farRight < farLeft));
-        }
-
-        bool checkCornersValid() const {
-            return br.n_elem == 2 && bl.n_elem == 2 && tr.n_elem == 2 && tl.n_elem == 2;
-        }
-
-        static Quad getBoundingBox(const std::vector<T>& points) {
-            // Check for
-            if (points.size() <= 0) {
-                throw std::domain_error("Request made for bounding box for empty list of points!");
-            }
-
-            Scalar min_x = points[0](0);
-            Scalar max_x = points[0](0);
-            Scalar min_y = points[0](1);
-            Scalar max_y = points[0](1);
-            for (uint i = 1; i < points.size(); i++) {
-                auto& p = points[i];
-                max_x   = std::max(max_x, p(0));
-                min_x   = std::min(min_x, p(0));
-                max_y   = std::max(max_y, p(1));
-                min_y   = std::min(min_y, p(1));
-            }
-            return Quad(T({min_x, min_y}), T({min_x, max_y}), T({max_x, max_y}), T({max_x, min_y}));
-        }
-
-    private:
-        T bl;  //! @variable The bottom-left of the Quad.
-        T br;  //! @variable The bottom-right of the Quad.
-        T tr;  //! @variable The top-right of the Quad.
-        T tl;  //! @variable The top-left of the Quad.
-
-        // //! @brief output stream operator.
-        // template <typename U>
-        // friend std::ostream& operator<<(std::ostream& output, const Quad<U>& quad);
-
-        // //! @brief output stream operator for a vector of goals.
-        // template <typename U>
-        // friend std::ostream& operator<<(std::ostream& output, const std::vector<Quad<U>>& quads);
-    };
-
-    // *****************
-    // * EIGEN VERSION *
-    // *****************
-
     template <typename Scalar, int R, int C>
-    class Quad<typename Eigen::Matrix<Scalar, R, C>> {
+    class Quad {
     public:
         using T = typename Eigen::Matrix<Scalar, R, C>;
         Quad() : bl(T::Zero()), br(T::Zero()), tr(T::Zero()), tl(T::Zero()) {}
@@ -398,19 +202,19 @@ namespace utility::math::geometry {
         T tl;  //! @variable The top-left of the Quad.
 
         //! @brief output stream operator.
-        template <typename U>
-        friend std::ostream& operator<<(std::ostream& output, const Quad<U>& quad);
+        template <typename Scalar_, int R_, int C_>
+        friend std::ostream& operator<<(std::ostream& output, const Quad<Scalar_, R_, C_>& quad);
 
         //! @brief output stream operator for a vector of goals.
-        template <typename U>
-        friend std::ostream& operator<<(std::ostream& output, const std::vector<Quad<U>>& quads);
+        template <typename Scalar_, int R_, int C_>
+        friend std::ostream& operator<<(std::ostream& output, const std::vector<Quad<Scalar_, R_, C_>>& quads);
     };
 
 
     /// @brief Stream insertion operator for a single Quad.
     /// @relates Quad
-    template <typename T>
-    inline std::ostream& operator<<(std::ostream& output, const Quad<T>& quad) {
+    template <typename Scalar, int R, int C>
+    inline std::ostream& operator<<(std::ostream& output, const Quad<Scalar, R, C>& quad) {
         output << "(" << quad.getBottomLeft().x() << ", " << quad.getBottomLeft().y() << ") (" << quad.getTopLeft().x()
                << ", " << quad.getTopLeft().y() << ") (" << quad.getTopRight().x() << ", " << quad.getTopRight().y()
                << ") (" << quad.getBottomRight().x() << ", " << quad.getBottomRight().y() << ")";
@@ -420,8 +224,8 @@ namespace utility::math::geometry {
 
     /// @brief Stream insertion operator for a std::vector of Quads.
     /// @relates Quad
-    template <typename T>
-    inline std::ostream& operator<<(std::ostream& output, const std::vector<Quad<T>>& quads) {
+    template <typename Scalar, int R, int C>
+    inline std::ostream& operator<<(std::ostream& output, const std::vector<Quad<Scalar, R, C>>& quads) {
         output << "[";
 
         for (const auto& quad : quads)
