+++ shared/utility/nusight/NUhelpers.hpp
@@ -20,7 +20,6 @@
 #ifndef NUHELPERS_HPP
 #define NUHELPERS_HPP
 
-#include <armadillo>
 #include <nuclear>
 #include <utility>
