+++ shared/utility/support/yaml_expression.hpp
@@ -21,7 +21,6 @@
 #define UTILITY_SUPPORT_yaml_expression_HPP
 
 #include <Eigen/Core>
-#include <armadillo>
 #include <fmt/format.h>
 #include <limits>
 #include <system_error>
@@ -268,108 +267,6 @@ namespace utility::support {
 
             return matrix;
         }
-
-        ////
-        //// ARMADILLO
-        ////
-
-        // Handle fixed-sized matrices.
-        template <typename T, std::enable_if_t<((!T::is_col) && (!T::is_row))>* = nullptr>
-        operator T() const {
-
-            // value : [[a, b], [c, d]]
-            const size_t rows = node.size();
-            const size_t cols = node[0].size();
-
-            // Count the columns on every row.
-            for (const auto& row : node) {
-                if (row.size() != cols) {
-                    throw std::out_of_range(
-                        fmt::format("Inconsistent number of cols in matrix (cols: {} vs {}).", row.size(), cols));
-                }
-            }
-
-            // Validate row and column sizes.
-            if ((rows != T::n_rows) || (cols != T::n_cols)) {
-                throw std::out_of_range(
-                    fmt::format("Rows and columns in YAML matrix do not align with output matrix. "
-                                "(rows: {} vs {}, cols: {} vs {}).",
-                                rows,
-                                T::n_rows,
-                                cols,
-                                T::n_cols));
-            }
-
-            T matrix;
-
-            try {
-                for (size_t row = 0; row < rows; row++) {
-                    for (size_t col = 0; col < cols; col++) {
-                        matrix(row, col) = parse_math_string<typename T::elem_type>(node[row][col].as<std::string>());
-                    }
-                }
-            }
-            catch (const std::invalid_argument& ex) {
-                throw std::invalid_argument(fmt::format("Unable to convert node to arithmetic type.\n{}", ex.what()));
-            }
-
-            return matrix;
-        }
-
-        // Handle fixed-sized column vectors.
-        template <typename T, std::enable_if_t<((T::is_col) && (!T::is_row))>* = nullptr>
-        operator T() const {
-
-            // Validate row size.
-            if (node.size() != T::n_rows) {
-                throw std::out_of_range(
-                    fmt::format("Rows in YAML column vector do not align with output vector. "
-                                "(rows: {} vs {}).",
-                                node.size(),
-                                T::n_rows));
-            }
-
-            T matrix;
-
-            try {
-                for (size_t i = 0; i < node.size(); i++) {
-                    matrix(i) = parse_math_string<typename T::elem_type>(node[i].as<std::string>());
-                }
-            }
-            catch (const std::invalid_argument& ex) {
-                throw std::invalid_argument(fmt::format("Unable to convert node to arithmetic type.\n{}", ex.what()));
-            }
-
-            return matrix;
-        }
-
-        // Handle fixed-sized row vectors.
-        template <typename T, std::enable_if_t<((!T::is_col) && (T::is_row))>* = nullptr>
-        operator T() const {
-
-            // Validate col size.
-            if (node.size() != T::n_cols) {
-                throw std::out_of_range(
-                    fmt::format("Columns in YAML row vector do not align with output vector. "
-                                "(cols: {} vs {}).",
-                                node.size(),
-                                T::n_cols));
-            }
-
-            T matrix;
-
-            try {
-                for (size_t i = 0; i < node.size(); i++) {
-                    matrix(i) = parse_math_string<typename T::elem_type>(node[i].as<std::string>());
-                }
-            }
-            catch (const std::invalid_argument& ex) {
-                throw std::invalid_argument(fmt::format("Unable to convert node to arithmetic type.\n{}", ex.what()));
-            }
-
-            return matrix;
-        }
-
 
     private:
         YAML::Node node;
